refactored repositories

+++ app/src/main/java/com/flydrop2p/flydrop2p/domain/model/Profile.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Profile(
-    val id: Long = -1,
+    val id: Int = -1,
     val username: String = android.os.Build.MODEL.toString()
 ) {
     override fun equals(other: Any?): Boolean = other is Profile && other.id == this.id
